% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_time.R
\name{adjust_time}
\alias{adjust_time}
\title{Generate climate dataset}
\usage{
adjust_time(hourly_data, tz = "America/Cayenne")
}
\arguments{
\item{hourly_data}{data.frame. Hourly or half-hourly data with a 'time'
column in UTC to be adjusted to the given time zone.}

\item{tz}{char. Time zone. Can be obtained from the coordinates with
\code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}.}
}
\value{
The same data frame with adjusted time.
}
\description{
Description
}
\examples{
adjust_time(data.frame(time = lubridate::as_datetime("2015-01-01 16:30:00")))

}
