% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.troll.R
\name{autoplot.troll}
\alias{autoplot.troll}
\alias{autoplot,trollsim-method}
\title{Plot TROLL outputs}
\usage{
\S4method{autoplot}{trollsim}(
  object,
  what = "ecosystem",
  variables = NULL,
  selected_species = "total"
)

\S4method{autoplot}{trollstack}(
  object,
  what = "ecosystem",
  variables = NULL,
  selected_species = "total"
)
}
\arguments{
\item{object}{TROLL simulation or stack.}

\item{what}{char. What to plot: "final pattern" or "ecosystem".}

\item{variables}{char. Wich variable(s) to plot: full simulation: "species",
"abu10", "abu30", "abund", "agb", "ba", "ba10", "gpp", "npp", "rday",
"rnight", or "rstem"; reduced simulation: "N", "N10", "N30", "BA10", "NPP",
"GPP", "AGB"}

\item{selected_species}{char. Which species to plot: "species name" or "total".}
}
\value{
ggplot2 object
}
\description{
This is a method to plot TROLL simulations or stacks, including either
temporal trajectories of whole ecosystem or species metrics, or the final
pattern observed in the forest community. Metrics includes abundances of
individuals above 1cm (abund & N), above 10cm (abu10 & N10), and above 30cm
(abu30 & N30), aboveground biomass (agb & AGB), basal area of individuals
above 1cm (ba, BA), and above 10cm (ba10, BA10), gross primary production
(gpp, GPP), net primary production (npp, NPP), and respiration of day (rday),
night (rnight) and stem (rstem).
}
\examples{
\dontrun{
autoplot(sim)
}

}
