% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.trollstack.R
\name{autoplot.trollstack}
\alias{autoplot.trollstack}
\alias{autoplot,trollstackfull-method}
\alias{autoplot,trollstackreduced-method}
\alias{autoplot,trollstackabc-method}
\title{Function to plot TROLL stack}
\usage{
\S4method{autoplot}{trollstackfull}(
  object,
  what = "ecosystem",
  variables = c("abu10", "abu30", "abund", "agb", "ba", "ba10", "gpp", "litterfall",
    "npp", "rday", "rnight", "rstem"),
  species = "total"
)

\S4method{autoplot}{trollstackreduced}(
  object,
  what = "ecosystem",
  variables = c("N", "N10", "N30", "BA10", "NPP", "GPP", "AGB"),
  species = "total"
)

\S4method{autoplot}{trollstackabc}(object, what)
}
\arguments{
\item{object}{trollsim}

\item{what}{char. "final pattern" or "ecosystem"}

\item{variables}{char. full outputs: "species", "abu10", "abu30", "abund",
"agb", "ba", "ba10", "gpp", "litterfall", "npp", "rday", "rnight", or
"rstem"; reduced outputs: "N", "N10", "N30", "BA10", "NPP", "GPP", "AGB"}

\item{species}{char. species name or total}

\item{...}{unused argument}
}
\value{
ggplot2 object
}
\description{
Function to plot TROLL stack
}
\examples{
NA

}
