% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build TROLL}
\usage{
build(src = getOption("RconTroll.src"), app = getOption("RconTroll.app"),
  path = getOption("RconTroll.path"), optimisation = NULL, verbose = TRUE)
}
\arguments{
\item{src}{char. path to src file}

\item{app}{char. path to troll app (e.g. TROLL.out)}

\item{path}{char. working directory}

\item{optimisation}{char. optimisation level to build cpp with g++ (default NULL)}

\item{verbose}{logical. activate outputs in console}
}
\value{
build the TROLL app
}
\description{
Function to build TROLL code
}
\examples{

}
