% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{compareplot}
\alias{compareplot}
\title{Compare plot}
\usage{
compareplot(sim, output = "agb", legend = TRUE, reduce = 12)
}
\arguments{
\item{sim}{list of TROLL.output. TROLL model outputs to compare}

\item{output}{char. Output to compare (abu10, abu30, agb and gpp available)}

\item{legend}{logical. add legend}

\item{reduce}{int. factor to reduce number of outputs points (default 12 months reduced to one year)}
}
\value{
ggplot2 object
}
\description{
function to compare
}
\examples{

}
