% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSim.R
\name{compareSim}
\alias{compareSim}
\title{Compare parameters of two or more TROLL simulations}
\usage{
compareSim(
  sim,
  ...,
  rowcols = TRUE,
  duration = TRUE,
  dimension = TRUE,
  species = FALSE,
  stopiffalse = TRUE,
  showwarning = FALSE
)
}
\arguments{
\item{sim}{TROLL.sim object}

\item{...}{TROLL.sim objects}

\item{rowcols}{logical. If TRUE, number of rows and columns of simulations are compared}

\item{duration}{logical. If TRUE, number of iterations and iterations per year are compared}

\item{dimension}{logical. If TRUE, cell dimensions (NH, NV) of TRoLL are compared}

\item{species}{logical. If TRUE, species number is compared}

\item{stopiffalse}{logical. If TRUE, an error will occur if the objects are not the same}

\item{showwarning}{logical. If TRUE, a warning will be given if objects are not the same. Only relevant when stopiffalse is not TRUE.}
}
\value{
logical
}
\description{
Function to compare whether two or more TROLLsim objects have been executed with the same set of general parameters. Please note: This function does not compare output.
}
\examples{
NA

}
