% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_troll.R
\name{compile_troll}
\alias{compile_troll}
\title{Compile troll}
\usage{
compile_troll(full = TRUE, abc = FALSE, forest = FALSE, random = TRUE)
}
\arguments{
\item{full}{bool. TROLL with full outputs, if not reduced outputs (default
TRUE)}

\item{abc}{bool. TROLL with abc outputs, forcing reduced outputs (default
FALSE)}

\item{forest}{bool. TROLL with forest input, if not start from an empty grid
(default FALSE)}

\item{random}{bool. TROLL with random outputs, if not the seed is fixed
(default TRUE)}
}
\value{
simulation outputs in the path folder
}
\description{
Compile troll
}
\examples{
\dontrun{
compile_troll()
}

}
