% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_sim.R
\name{date_sim}
\alias{date_sim}
\alias{date_sim,trollsim-method}
\title{Function to date outputs from a TROLL simulation based on a starting date.}
\usage{
date_sim(sim, date, ...)

\S4method{date_sim}{trollsim}(sim, date, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{date}{char. date as a string YYYY/MM/DD.}

\item{...}{unused argument.}
}
\value{
An S4 \linkS4class{trollsim} class object.
}
\description{
Function to date outputs from a TROLL simulation based on a starting date.
}
