% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-misc.R
\name{dewtovpd}
\alias{dewtovpd}
\title{dewtovpd}
\usage{
dewtovpd(temperature, dewpoint, pressure = 101325)
}
\arguments{
\item{temperature}{num. Temperature in Kelvin.}

\item{dewpoint}{num. Temperature in Kelvin.}

\item{pressure}{num.  Atmospheric pressure in Pa, default 1 atmosphere.}
}
\value{
Estimated vapour pressure in Pa
}
\description{
function to compute VPD from temperature, dewpoint temperature & surface
pressure
}
\details{
from Jones, H.G. 2014. Plants and microclimate: a quantitative approach to
environmental plant physiology. 3nd Edition., 2nd Edn. Cambridge University
Press, Cambridge. 428 p.
}
\examples{
dewtovpd(27 + 273.15, 25 + 273.15)

}
