% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era.R
\name{download_era}
\alias{download_era}
\title{Download ERA5 data for TROLL}
\usage{
download_era(
  user,
  variables = c("u10", "v10", "d2m", "t2m", "ssr", "sp", "tp"),
  years = 2020,
  months = 1,
  days = 1:31,
  hours = 0:23,
  target = NULL,
  area = c(5.77, -54.6, 2.11, -51.63),
  format = "netcdf",
  request = FALSE,
  transfer = TRUE,
  path = getwd(),
  verbose = FALSE,
  time_out = 60 * 60
)
}
\arguments{
\item{user}{}

\item{variables}{}

\item{years}{}

\item{months}{}

\item{days}{}

\item{hours}{}

\item{target}{}

\item{area}{}

\item{format}{}

\item{request}{}

\item{transfer}{}

\item{path}{}

\item{verbose}{}

\item{time_out}{}
}
\value{
Write files in the defined path or output the request as a list if
\code{request=FALSE}.
}
\description{
Description.
}
