% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-misc.R
\name{esat}
\alias{esat}
\title{esat}
\usage{
esat(temperature, pressure = 101325)
}
\arguments{
\item{temperature}{num. Temperature in celsius degrees.}

\item{pressure}{num.  Atmospheric pressure in Pa, default 1 atmosphere.}
}
\value{
Estimated vapour pressure in Pa
}
\description{
Compute vapour pressure from temperature & surface pressure
}
\details{
from Jones, H.G. 2014. Plants and microclimate: a quantitative approach to
environmental plant physiology. 3nd Edition., 2nd Edn. Cambridge University
Press, Cambridge. 428 p.
}
\examples{
esat(27)

}
