% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_era.R
\name{extract_era}
\alias{extract_era}
\title{Extract ERA}
\usage{
extract_era(era5)
}
\arguments{
\item{era5}{str. Path to ERA5 land data in netCDF. See the corresponding
vignette \code{vignette("climate", package = "rcontroll")} to download
corresponding data from Copernicus in R.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with ERA5-Land data.
}
\description{
Description
}
\details{
Details on computations.
}
