% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_era.R
\name{extract_era_batch}
\alias{extract_era_batch}
\title{Extract ERA batch}
\usage{
extract_era_batch(era5_files, cores = NULL)
}
\arguments{
\item{era5_files}{vector. Vector of paths to ERA5 land data in netCDF. See
the corresponding vignette \code{vignette("climate", package =
  "rcontroll")} to download corresponding data from Copernicus in R.}

\item{cores}{int. Number of cores for parallelization, if NULL available
cores - 1 (default NULL).}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with ERA5-Land data..
}
\description{
Description
}
\details{
Details on computations.
}
