% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_era.R
\name{format_era}
\alias{format_era}
\title{format ERA}
\usage{
format_era(era5, tz = "America/Cayenne")
}
\arguments{
\item{era5}{data.frame. Data from ERA5-Land. See the corresponding vignette
\code{vignette("climate", package = "rcontroll")} to download corresponding
data from Copernicus in R.}

\item{tz}{num. Time zone. Can be obtained from the coordinates with
\code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with formatted ERA5-Land data for TROLL.
}
\description{
Description
}
\details{
Details on computations.
}
