% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dailyvar.R
\name{generate_dailyvar}
\alias{generate_dailyvar}
\title{Generate climate dataset}
\usage{
generate_dailyvar(
  hourly_data,
  daytime_start = 7,
  daytime_end = 19,
  tz = "America/Cayenne"
)
}
\arguments{
\item{hourly_data}{data.frame. Half-hourly climatic data formatted for TROLL.}

\item{daytime_start}{int. Daytime starting hour to compute nigh and day
variables (default 7).}

\item{daytime_end}{int. Daytime ending hour to compute nigh and day variables
(default 19).}

\item{tz}{num. Time zone. Can be obtained from the coordinates with
\code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with climatic half-hourly variation (temperature,
wind speed, shortwave irradiation and vapour pressure deficit).
}
\description{
Description
}
\details{
Details.
}
