% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chm.trollsim.R
\name{get_chm}
\alias{get_chm}
\alias{get_chm,trollsim-method}
\title{Extract canopy height model}
\usage{
get_chm(sim, method = "smoothed", ...)

\S4method{get_chm}{trollsim}(sim, method = "smoothed", ...)
}
\arguments{
\item{sim}{trollsim.}

\item{method}{char. method to extract the canopy height model from the point
cloud in las, either 'filled' (replacing NA by 0) or 'smoothed' (local
means, default value).}

\item{...}{unused argument.}
}
\value{
\code{\link[=data.frame]{data.frame()}}
}
\description{
\code{get_chm()} extract the canopy height model from \code{TROLL} outputs with lidar
option.
}
\seealso{
\code{\link[=trollsim]{trollsim()}}, \code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}
}
