% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest.trollsim.R
\name{get_forest}
\alias{get_forest}
\alias{get_forest,trollsim-method}
\title{Extract forest inventory}
\usage{
get_forest(sim, ...)

\S4method{get_forest}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim or trollstack.}

\item{...}{unused argument.}
}
\value{
\code{\link[=data.frame]{data.frame()}}
}
\description{
\code{get_forest()} extract the forest inventory from \code{TROLL} outputs.
}
\examples{

data("TROLLv3_output")
head(get_forest(TROLLv3_output))

}
\seealso{
\code{\link[=trollsim]{trollsim()}}, \code{\link[=trollstack]{trollstack()}}, \code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}
}
