% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.trollsim.R
\name{get_log}
\alias{get_log}
\alias{get_log,trollsim-method}
\title{Extract simulation log}
\usage{
get_log(sim, ...)

\S4method{get_log}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim or trollstack.}

\item{...}{unused argument.}
}
\value{
the log in the console
}
\description{
\code{get_forest()} extract the simulation log \code{TROLL} outputs.
}
\examples{

data("TROLLv3_output")
get_log(TROLLv3_output)

}
\seealso{
\code{\link[=trollsim]{trollsim()}}, \code{\link[=trollstack]{trollstack()}}, \code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}
}
