% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTROLL.R
\name{getTROLL}
\alias{getTROLL}
\title{get TROLL}
\usage{

  getTROLL(source = "https://raw.githubusercontent.com/TROLL-code/TROLL/master/main_v2.3.2.cpp",
  destination = getOption("RconTROLL.src"),
  app = getOption("RconTroll.app"))
}
\arguments{
\item{source}{char. link to the TROLL C++ source code online (default: GitHub
TROLL-code official repository)}

\item{destination}{char. path to the TROLL C++ source code locally (default:
package tools folder)}

\item{app}{char. (default: package tools folder)}
}
\value{
copy TROLL C++ code and compile it into your tools folder
}
\description{
Locally copy TROLL C++ code into your package installation and eventually
complie it into an executable (linux/UNIX only).
}
\examples{

\dontrun{
getTROLL(source = 
"https://raw.githubusercontent.com/TROLL-code/TROLL/master/main_v2.3.2.cpp", 
destination = getOption("RconTROLL.src"), app = getOption("RconTroll.app"))
}

}
