% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryFromOutput.R
\name{inventoryFromOutput}
\alias{inventoryFromOutput}
\title{Create inventory from output}
\usage{
inventoryFromOutput(x, forest = getOption("RconTroll.forest"),
  path = getOption("RconTroll.path"), overwrite = TRUE)
}
\arguments{
\item{x}{TROLLsim}

\item{forest}{char. name of the file to write}

\item{path}{char. path to the working directory}

\item{overwrite}{logical. allow to overwrite existing input file}
}
\value{
virtual forest file for the TROLL program (forest.txt)
}
\description{
Create an inventory from a TROLL model output
}
\examples{
NA

}
