% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.stack.R
\name{is.stack}
\alias{is.stack}
\title{Function to test if TROLL outputs is a stack.}
\usage{
is.stack(x)
}
\arguments{
\item{x}{object}
}
\value{
logical
}
\description{
Function to test if TROLL outputs is a stack.
}
\examples{
NA

}
