% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{Function to load TROLL output}
\usage{
load(name, path = getwd())
}
\arguments{
\item{name}{char. Name given to the model output}

\item{path}{char. Path where the model is saved}
}
\value{
an S4 \linkS4class{TROLLoutput} class object
}
\description{
Function to load TROLL output
}
\examples{

}
