% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stack.R
\name{load_stack}
\alias{load_stack}
\title{Function to load TROLL stack}
\usage{
load_stack(name, path, type)
}
\arguments{
\item{name}{char. name given to the stack output}

\item{path}{char. path where the stack is saved}

\item{type}{char. type of simulation (full, reduce or abc)}
}
\value{
an S4 \linkS4class{trollsim} class object
}
\description{
Function to load TROLL stack
}
\examples{

NA
}
