% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TOLLoutput.R
\docType{methods}
\name{plot.TROLLoutput}
\alias{plot.TROLLoutput}
\title{Function to plot TROLL outputs.}
\usage{
\S4method{plot}{TROLLoutput,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{TROLLoutput}

\item{y}{char. Output to plot, see details.}
}
\value{
Plot the output
}
\description{
Function to plot TROLL outputs.
}
\details{
Available plots:
\describe{
\item{defaults}{abund + agb + ba +height hist}
\item{abund}{
 \describe{
 \item{abund}{Number of stems/ha ~ Time (in years)}
 \item{abu10}{Number of trees with dbh > 10 cm (in stems/ha) ~ Time (in years)}
 \item{abu30}{Number of trees with dbh > 30 cm (in stems/ha) ~ Time (in years)}
 \item{allabund}{abund + abu10 + abu30}
 }
}
\item{agb}{Aboveground biomass (in tonnes/ha) ~ Time (in years)}
\item{ba}{
 \describe{
 \item{ba}{Basal area (in m2/ha) ~ Time (in years)}
 \item{ba10}{Basal area of trees with dbh > 10 cm (in stems/ha) ~ Time (in years)}
 \item{allba}{ba + ba10}
 }
}
\item{death}{Not implemented}
\item{disturbance}{
 \describe{
 \item{dist height}{disturbed tree height histogram}
 \item{dist dbh}{disturbed dbh histogram}
 \item{dist rank-abundance}{disturbed species rank-abundance histogram}
 }
}
\item{final pattern}{
 \describe{
 \item{heigh}{final tree height histogram}
 \item{dbh}{final dbh histogram}
 \item{rank-abundance}{final species rank-abundance histogram}
 \item{age}{final tree age distribution}
 \item{species}{final species spatial distribution map}
 }
}
\item{flux}{GPP, Autotrphic respiration, NPP ~ Time (in years)}
\item{gpp}{Total GPPLeaf (in MgC/ha) ~ Time (in years)}
\item{LAI}{Not implemented}
\item{litterfall}{Leaf litterfall per month (in Mg/ha/year dry mass) ~ Time (in years)}
\item{npp}{Total NPPLeaf (in MgC/ha) ~ Time (in years)}
\item{R}{Not implemented}
\item{Ra}{Autotrophic respiration (in MgC/ha) ~ Time (in years)}
\item{relabund}{Not implemented}
\item{vertd}{PAD (m2 per m3) ~ Height (in m)}
\item{wd}{
 \describe{
 \item{wd}{Total wood density (in g/cm3) ~ Time (in years)}
 \item{wd10}{Wood density of trees with dbh > 10 cm (in g/cm3) ~ Time (in years)}
 \item{wd30}{Wood density of trees with dbh > 30 cm (in g/cm3) ~ Time (in years)}
 \item{allwd}{wd + wd10 + wd30}
 }
}
}
}
\examples{

}
