% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TROLLsim.R
\docType{methods}
\name{plot.TROLLsim}
\alias{plot.TROLLsim}
\alias{plot,TROLLsim,missing-method}
\alias{plot,TROLLsim,TROLLsim-method}
\alias{plot,TROLLsimstack,missing-method}
\alias{plot,TROLLsimstack,TROLLsimstack-method}
\title{Plot TROLL simulations stack}
\usage{
\S4method{plot}{TROLLsim,missing}(x, y, what, ...)

\S4method{plot}{TROLLsim,TROLLsim}(x, y, what, ...)

\S4method{plot}{TROLLsimstack,missing}(x, y, what, ...)

\S4method{plot}{TROLLsimstack,TROLLsimstack}(x, y, what, ...)
}
\arguments{
\item{x}{TROLLsim or TROLLsimstack. Troll simulation stacked or not object to
plot}

\item{y}{Any, TROLLsim or TROLLsimstack. nothing, TROLL simulation or Troll
aggregated simulation stack to compare with}

\item{what}{char. ecosystem output to plot, see details}

\item{...}{other graphical parameters}
}
\value{
Plot the simulations
}
\description{
Plot TROLL simulations stack
}
\details{
Available plots:
\describe{
\item{agb}{above ground biomass}
\item{gpp}{gross primary productivity}
\item{abund, abu10, abu30}{abundances (total, above 10 and above 30 cm dbh)}
\item{ba, ba10}{basal area (total and above 10 cm dbh)}
\item{Rday, Rnight}{night and day respiration}
\item{diversity, distdiversity}{species diversity for different order of diversities
 (final and disturbed, require entropart)}
\item{rankabund, distrankabund}{species rank-abundance diagram(final and disturbed)}
\item{height, distheight}{tree height histogram (final and disturbed)}
\item{dbh, distdbh}{tree dbh histogram (final and disturbed)}
\item{age, distage}{tree age histogram (final and disturbed)}
\item{trait, disttrait}{functional trait density distribution, replace trait
by the name of the trait you want to plot (final and disturbed)}
}
}
\examples{
NA

}
