% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TROLLsimstack.R
\docType{methods}
\name{plot.TROLLsimstack}
\alias{plot.TROLLsimstack}
\alias{plot,TROLLsimstack,missing-method}
\alias{plot,TROLLsimstack,TROLLsimstack-method}
\title{Plot TROLL simulations stack}
\usage{
\S4method{plot}{TROLLsimstack,missing}(x, y, what, ggplot2 = FALSE,
  plotly = FALSE, ...)

\S4method{plot}{TROLLsimstack,TROLLsimstack}(x, y, what, ggplot2 = FALSE,
  plotly = FALSE, ...)
}
\arguments{
\item{x}{TROLLsimstack. Troll simulation stack object to plot}

\item{y}{TROLLsimstack. nothing or Troll aggregated simulation stack to
compare with}

\item{what}{char. ecosystem output to plot, see details}

\item{ggplot2}{logical. creates ggplot graph}

\item{plotly}{logical. use plotly library for interactive plots with ggplot}

\item{...}{other graphical parameters}
}
\value{
Plot the simulations
}
\description{
Plot TROLL simulations stack
}
\details{
Available plots:
\describe{
\item{agb}{above ground biomass}
}
}
\examples{

}
