% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecosystem.R
\name{plot_ecosystem}
\alias{plot_ecosystem}
\title{Function to plot ecosystem metrics from trollsim}
\usage{
plot_ecosystem(sim)
}
\arguments{
\item{sim}{trollsim. trollsim to be plotted}
}
\value{
a ggplot object
}
\description{
Function to plot ecosystem metrics from trollsim
}
\examples{

NA
}
