% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_era.R
\name{prepare_era}
\alias{prepare_era}
\title{Prepare ERA}
\usage{
prepare_era(era5, tz = "America/Cayenne")
}
\arguments{
\item{era5}{str. Path to ERA5 land data in netCDF. See the corresponding
vignette \code{vignette("climate", package = "rcontroll")} to download
corresponding data from Copernicus in R.}

\item{tz}{num. Time zone. Can be obtained from the coordinates with
\code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with...
}
\description{
Description
}
\details{
Details on computations.
}
