% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_era_batch.R
\name{prepare_era_batch}
\alias{prepare_era_batch}
\title{Prepare ERA batch}
\usage{
prepare_era_batch(era5_files, cores = NULL, tz = "America/Cayenne")
}
\arguments{
\item{era5_files}{vector. Vector of paths to ERA5 land data in netCDF. See
the corresponding vignette \code{vignette("climate", package =
  "rcontroll")} to download corresponding data from Copernicus in R.}

\item{cores}{int. Number of cores for parallelization, if NULL available
cores - 1 (default NULL).}

\item{tz}{num. Time zone. Can be obtained from the coordinates with
\code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with...
}
\description{
Description
}
\details{
Details on computations.
}
