% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TROLLsim.R
\name{print.TROLLsim}
\alias{print.TROLLsim}
\alias{print,TROLLsim-method}
\title{Function to print or show TROLL outputs.}
\usage{
\S4method{print}{TROLLsim}(x, ...)
}
\arguments{
\item{x}{TROLLsim or TROLLsim}

\item{...}{unused argument}
}
\value{
Print or show in console
}
\description{
Function to print or show TROLL outputs.
}
\examples{
NA

}
