% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TROLLsimstack.R
\docType{methods}
\name{print.TROLLsimstack}
\alias{print.TROLLsimstack}
\title{Function to print TROLL outputs.}
\usage{
\S4method{print}{TROLLsimstack}(x, ...)
}
\arguments{
\item{x}{TROLLsimstack}
}
\value{
Print in console
}
\description{
Function to print TROLL outputs.
}
\examples{

}
