% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_era.R
\name{request_era}
\alias{request_era}
\title{Request ERA5 data for TROLL}
\usage{
request_era(
  variables = c("u10", "v10", "d2m", "t2m", "ssr", "sp", "tp"),
  year = 2004,
  months = 1:12,
  days = 1:31,
  hours = 0:23,
  prefix = "ERA5land_Paracou",
  xutm = 5.26724134423233,
  yutm = -52.924368025558,
  format = "netcdf"
)
}
\arguments{
\item{variables}{char. Description.}

\item{year}{int. Description.}

\item{months}{int. Description.}

\item{days}{int. Description.}

\item{hours}{int. Description.}

\item{prefix}{char. Description.}

\item{xutm}{int. Description.}

\item{yutm}{int. Description.}

\item{format}{char. Description.}
}
\value{
ERA request for \code{ecmwfr} as a list of attributes.
}
\description{
Description.
}
