% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_era_batch.R
\name{request_era_batch}
\alias{request_era_batch}
\title{Request ERA5 batch data for TROLL}
\usage{
request_era_batch(
  variables = c("u10", "v10", "d2m", "t2m", "ssr", "sp", "tp"),
  years = 2020,
  months = 1,
  days = 1:31,
  hours = 0:23,
  target = NULL,
  area = c(5.77, -54.6, 2.11, -51.63),
  format = "netcdf"
)
}
\arguments{
\item{request}{}
}
\value{
List of ERA requests for \code{ecmwfr} as a list of attributes.
}
\description{
Description.
}
