% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-misc.R
\name{rh_to_vpd}
\alias{rh_to_vpd}
\title{rh_to_vpd}
\usage{
rh_to_vpd(humidity, temperature, pressure = 101)
}
\arguments{
\item{humidity}{num. Relative humidity in \%.}

\item{temperature}{num. Temperature in celsius degrees.}

\item{pressure}{num.  Atmospheric pressure in Pa, default 1 atmosphere.}
}
\value{
Estimated vapour pressure in Pa
}
\description{
Compute vapour pressure deficit from relative humidity
}
\details{
from Jones, H.G. 2014. Plants and microclimate: a quantitative approach to
environmental plant physiology. 3nd Edition., 2nd Edn. Cambridge University
Press, Cambridge. 428 p.
}
\examples{
rh_to_vpd(90, 27)

}
