% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run TROLL}
\usage{
run(path, name = getOption("RconTROLL.name"),
  app = getOption("RconTroll.app"), input = getOption("RconTroll.init"),
  overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{char. working directory}

\item{name}{char. name of the model}

\item{app}{char. path to troll app (e.g. TROLL.out)}

\item{input}{char. input file}

\item{overwrite}{logical. allow to overwrite existing outputs files}

\item{verbose}{logical. allow output in console}
}
\value{
model output files in the path folder
}
\description{
Function to run TROLL application
}
\examples{

\dontrun{
getTROLL()
init(path = system.file("tools", package = 'RconTROLL'),
     input = "init.txt")
run(path = system.file("tools", package = 'RconTROLL'),
    name = "test", app = getOption("RconTroll.app"),
    input = system.file("tools", "init.txt", package = 'RconTROLL'))
}

}
