% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Stack}
\usage{
stack(
  name = NULL,
  simulations,
  path = NULL,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  verbose = TRUE,
  cores = NULL,
  overwrite = TRUE,
  thin = NULL
)
}
\arguments{
\item{name}{char. stack name (if NULL timestamp)}

\item{simulations}{char. simulations names}

\item{path}{char. path to the stack (tmp if NULL)}

\item{global}{df. global parameters, add a simulation column for repeated
simulations}

\item{species}{df. species parameters, add a simulation column for repeated
simulations}

\item{climate}{df. climate parameters, add a simulation column for repeated
simulations}

\item{daily}{df. daily variation parameters, add a simulation column for
repeated simulations}

\item{forest}{df. TROLL with forest input, if null start from an empty grid,
add a simulation column for repeated simulations (default NULL)}

\item{verbose}{bool. show TROLL outputs in the console}

\item{cores}{int. number of cores for parallelization, if NULL available
cores - 1 (default NULL)}

\item{overwrite}{bool. overwrite previous outputs}

\item{thin}{int. vector of integers corresponding to iterations to be kept
(default NULL)}
}
\value{
trollstack
}
\description{
Run a TROLL stack
}
\examples{
\dontrun{
example
}

}
