% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.TROLLsim.R
\docType{methods}
\name{stack.TROLLsim}
\alias{stack.TROLLsim}
\alias{stack,TROLLsim-method}
\alias{stack,list-method}
\title{Create TROLL simulations stack}
\usage{
\S4method{stack}{TROLLsim}(x, ..., layers = NULL)

\S4method{stack}{list}(x, bands = NULL)
}
\arguments{
\item{x}{TROLLsim. Troll simulation object}

\item{...}{other TROLL simulations}

\item{layers}{unused argument}

\item{bands}{unused arguments}
}
\value{
TROLL simulation stack
}
\description{
Simulation stack is a collection of TROLL simulations created from several or
a list off TROLL simulations object.
}
\examples{
NA

}
