% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TROLLsim.R
\docType{methods}
\name{summary.TROLLsim}
\alias{summary.TROLLsim}
\title{Function to provide summary of TROLL outputs}
\usage{
\S4method{summary}{TROLLsim}(object, ...)
}
\arguments{
\item{x}{TROLLsim}
}
\value{
Print in console
}
\description{
Function to provide summary of TROLL outputs
}
\examples{

}
