% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{Run a \code{TROLL} simulation}
\usage{
troll(
  name = NULL,
  path = NULL,
  global,
  species,
  climate,
  daily,
  pedology,
  forest = NULL,
  soil = NULL,
  lidar = NULL,
  load = TRUE,
  verbose = TRUE,
  overwrite = TRUE,
  date = NULL,
  thin = NULL
)
}
\arguments{
\item{name}{char. Model name (if NULL the timestamp will be used).}

\item{path}{char. Path to save the simulation outputs, the default is null
corresponding to a simulation in memory without saved intermediary files
(based on temporary files from \link{option.rcontroll}).}

\item{global}{df. Global parameters (e.g. \link{TROLLv4_input} or using
\code{\link[=generate_parameters]{generate_parameters()}}).}

\item{species}{df. Species parameters (e.g. \link{TROLLv4_species}).}

\item{climate}{df. Climate parameters (e.g. \link{TROLLv3_climate}).}

\item{daily}{df. Daily variation parameters (e.g. \link{TROLLv4_dailyvar}).}

\item{pedology}{df. pedology parameters (e.g. \link{TROLLv4_pedology}).}

\item{forest}{df. TROLL with forest input, if null starts from an empty grid
(default NULL) (e.g. using \link{TROLLv4_output} with \code{\link[=get_forest]{get_forest()}}).}

\item{soil}{df. TROLL with soil input, if null starts from an empty grid
(default NULL) (e.g. using \link{TROLLv4_output} with \code{\link[=get_soil]{get_soil()}}).}

\item{lidar}{df. Lidar simulation parameters (e.g. using \code{\link[=generate_lidar]{generate_lidar()}}),
if null not computed (default NULL).}

\item{load}{bool. TROLL outputs are loaded in R memory, if not only the path
and name of the stack of simulations is kept in the resulting
\code{\link[=trollstack]{trollstack()}} object but the content can be accessed later using the
\code{\link[=load_sim]{load_sim()}} method.}

\item{verbose}{bool. Show TROLL log in the console.}

\item{overwrite}{bool. Overwrite previous outputs folder and files.}

\item{date}{char. Starting date as YYYY/MM/DD, default NULL will result in
non-dated outputs.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce output size, default is NULL and corresponds to no thinning.}
}
\value{
A \code{\link[=trollsim]{trollsim()}} object.
}
\description{
\code{troll()} run a \code{TROLL} simulation. The minimal set of input files required
for a \code{TROLL} run include (i) climate data for the focal location (\code{climate}
and \code{daily}), (ii) soil data for the focal location (\code{pedology}), (iii)
functional traits for the list of species at the focal location (\code{species}),
and (iv) global parameters (\code{global}), i.e. parameters that do not depend on
species identity.
}
\examples{
\donttest{
data("TROLLv4_input")
data("TROLLv4_species")
data("TROLLv4_climate")
data("TROLLv4_dailyvar")
data("TROLLv4_pedology")
sim <- troll(
  name = "test",
  path = getwd(),
  global = generate_parameters(nbiter = 10),
  species = TROLLv4_species,
  climate = TROLLv4_climate,
  daily = TROLLv4_dailyvar,
  pedology = TROLLv4_pedology,
  load = TRUE,
  date = "2004/01/01"
)
}

}
\seealso{
\code{\link[=stack]{stack()}}
}
