% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLversion.R
\name{TROLL.version}
\alias{TROLL.version}
\title{\code{TROLL} version}
\usage{
TROLL.version()
}
\value{
\code{TROLL} version in console.
}
\description{
\code{TROLL.version()} prints \code{TROLL} version.
}
\examples{

TROLL.version()

}
\seealso{
\link{option.rcontroll}
}
