% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trollCpp}
\alias{trollCpp}
\title{TROLL simulator}
\usage{
trollCpp(
  global_file,
  climate_file,
  species_file,
  day_file,
  lidar_file,
  forest_file,
  output_file
)
}
\arguments{
\item{global_file}{char. Path to the global parameters file.}

\item{climate_file}{char. Path to the climate file.}

\item{species_file}{char. Path to the species file.}

\item{day_file}{char. Path to the daytime file.}

\item{lidar_file}{char. Path to the lidar file.}

\item{forest_file}{char. Path to the forest file.}

\item{output_file}{char. Path to the output folder.}
}
\value{
Void with outputs files written in the defined folder.
}
\description{
Wrapper of the TROLL C++ simulator with Rcpp.
}
\examples{
\dontrun{
trollCpp(
  global_file = "test/test_input_global.txt",
  climate_file = "test/test_input_climate.txt",
  species_file = "test/test_input_species.txt",
  day_file = "test/test_input_daily.txt",
  lidar_file = "",
  forest_file = "",
  output_file = "test"
)
}

}
