% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsimabc.R
\docType{class}
\name{trollsimabc-class}
\alias{trollsimabc-class}
\title{An S4 class to represent TROLL simulations}
\description{
This is an S4 class to represent TROLL simulations with abc outputs.
}
\section{Slots}{

\describe{
\item{\code{abc_outputs}}{list.. ABC metrics}
}}

