% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsimfull.R
\docType{class}
\name{trollsimfull-class}
\alias{trollsimfull-class}
\title{An S4 class to represent TROLL simulations}
\description{
This is an S4 class to represent TROLL simulations with full outputs.
}
\section{Slots}{

\describe{
\item{\code{species_outputs}}{df. species and ecosystem metrics}
}}

