% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsimfull.R
\name{trollsimfull}
\alias{trollsimfull}
\title{An S4 class to represent TROLL simulations}
\usage{
trollsimfull(
  name = character(),
  path = character(),
  species_outputs = data.frame(),
  inputs = list(),
  log = character()
)
}
\arguments{
\item{name}{char. model name}

\item{path}{char. path to the model}

\item{species_outputs}{df. species and ecosystem metrics}

\item{inputs}{list. model inputs (general, species, climate, daily, forest)}

\item{log}{chr. model log}
}
\description{
This is an S4 class to represent TROLL simulations.
}
