% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsimreduced.R
\docType{class}
\name{trollsimreduced-class}
\alias{trollsimreduced-class}
\title{An S4 class to represent TROLL simulations}
\description{
This is an S4 class to represent TROLL simulations with reduced outputs.
}
\section{Slots}{

\describe{
\item{\code{reduced_outputs}}{df. ecosystem metrics}
}}

