% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLsimstack.R
\docType{class}
\name{TROLLsimstack-class}
\alias{TROLLsimstack-class}
\title{An S4 class to represent TROLL outputs}
\description{
This is an S4 class to represent TROLL outputs in stacked format (stack of TROLLsim objects)
}
\section{Slots}{

\describe{
\item{\code{structured}}{logical. indicates whether stacked simulations have the same dimensions or not}

\item{\code{compressed}}{logical. indicates whether stacked simulations are saved in compressed format (mean, sd)}

\item{\code{layers}}{list. TROLLsim objects}
}}

