% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollstackabc.R
\name{trollstackabc}
\alias{trollstackabc}
\title{An S4 class to represent TROLL stack}
\usage{
trollstackabc(
  name = character(),
  path = character(),
  parameters = numeric(),
  inputs = list(),
  log = character(),
  final_pattern = data.frame(),
  reduced_outputs = data.frame(),
  abc_outputs = list()
)
}
\arguments{
\item{name}{char. simulation name}

\item{path}{char. path to the simulation}

\item{parameters}{numeric. parameters of the simulation (general inputs, random, forest)}

\item{inputs}{list. simulation inputs (species, climate, daily, forest)}

\item{log}{chr. simulation log}

\item{final_pattern}{df. simulation final forest}

\item{reduced_outputs}{df. ecosystem metrics}

\item{abc_outputs}{list. abc metrics}
}
\description{
This is an S4 class to represent TROLL stack
}
