% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollstackreduced.R
\docType{class}
\name{trollstackreduced-class}
\alias{trollstackreduced-class}
\title{An S4 class to represent TROLL stack}
\description{
This is an S4 class to represent TROLL stack with reduced outputs.
}
