% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollstackreduced.R
\name{trollstackreduced}
\alias{trollstackreduced}
\title{An S4 class to represent TROLL stack}
\usage{
trollstackreduced(
  name = character(),
  path = character(),
  parameters = numeric(),
  inputs = list(),
  log = character(),
  final_pattern = data.frame(),
  reduced_outputs = data.frame()
)
}
\arguments{
\item{name}{char. simulation name}

\item{path}{char. path to the simulation}

\item{parameters}{numeric. parameters of the simulation (general inputs, random, forest)}

\item{inputs}{list. simulation inputs (species, climate, daily, forest)}

\item{log}{chr. simulation log}

\item{final_pattern}{df. simulation final forest}

\item{reduced_outputs}{df. ecosystem metrics}
}
\description{
This is an S4 class to represent TROLL stack.
}
