% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_input.R
\docType{data}
\name{TROLLv3_input}
\alias{TROLLv3_input}
\title{\code{TROLL} global parameters}
\format{
A data frame with 61 rows and 3 variables: \describe{
\item{param}{global parameter} \item{value}{value of the parameter}
\item{description}{description of the parameter} }
}
\usage{
TROLLv3_input
}
\description{
Global parameters definition used by \code{TROLL} model.
}
\seealso{
\code{\link[=generate_parameters]{generate_parameters()}}
}
\keyword{datasets}
