% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_output.R
\docType{data}
\name{TROLLv3_output}
\alias{TROLLv3_output}
\title{\code{TROLL} output}
\format{
A \code{\link[=trollsim]{trollsim()}} object.
}
\usage{
TROLLv3_output
}
\description{
\code{TROLL} outputs from a 100-year simulation on a 100x100 grid with the other
default parameters and using TROLLv3_species, TROLLv3_climatedaytime12, and
TROLLv3_daytimevar for use in tests and examples. Ecosystem level output has
been thinned and species output has been removed to save disk space.
}
\seealso{
\code{\link[=TROLLv3_species]{TROLLv3_species()}}, \code{\link[=TROLLv3_climatedaytime12]{TROLLv3_climatedaytime12()}},
\code{\link[=TROLLv3_daytimevar]{TROLLv3_daytimevar()}}, \code{\link[=troll]{troll()}}
}
\keyword{datasets}
