% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_pointcloud.R
\docType{data}
\name{TROLLv3_pointcloud}
\alias{TROLLv3_pointcloud}
\title{TROLL lidar parameters}
\format{
A data frame with 5 rows and 3 variables: \describe{
  \item{param}{global parameter} \item{value}{value of the parameter}
  \item{description}{description of the parameter} }
}
\usage{
TROLLv3_pointcloud
}
\description{
Lidar parameters definition used by TROLL model for lidar simulations.
}
\keyword{datasets}
