% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_species.R
\docType{data}
\name{TROLLv3_species}
\alias{TROLLv3_species}
\title{\code{TROLL} species parameters}
\format{
A data frame with 45 rows and 12 variables: \describe{
\item{s_name}{Species name genus_species} \item{s_LMA}{leaf mass per area}
\item{s_Nmass}{leaf nitrogen mass} \item{s_Pmass}{leaf phosphorus mass}
\item{s_wsg}{wood specific gravity} \item{s_dbhmax}{maximum diameter}
\item{s_hmax}{maximum height} \item{s_ah}{height-diameter allometry
coefficient} \item{s_CD_a}{crown depth allometry intercept}
\item{s_CD_b}{crown depth allometry slope} \item{s_CR_a}{crown radius
allometry intercept} \item{s_CR_b}{crown radius allometry slope} }
}
\usage{
TROLLv3_species
}
\description{
Functional traits used by \code{TROLL} model for 70 species in Central Africa
gathered by Nicholas Russo (Harvard).
}
\seealso{
\code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}
}
\keyword{datasets}
