% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv4_climate.R
\docType{data}
\name{TROLLv4_climate}
\alias{TROLLv4_climate}
\title{TROLL climate parameters over days}
\format{
A data frame with 730 rows and 12 variables: \describe{
\item{NightTemperature}{night temperature (degree C)}
\item{Rainfall}{rainfall (mm)}}
}
\usage{
TROLLv4_climate
}
\description{
Climate parameters used by TROLL model over days with daily means from 2014 &
2015 in Paracou.
}
\keyword{datasets}
