% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv4_dailyvar.R
\docType{data}
\name{TROLLv4_dailyvar}
\alias{TROLLv4_dailyvar}
\title{TROLL daytime variation parameters}
\format{
A data frame with 24 rows and 5 variables: \describe{
  \item{DayJulian}{Julian day since 2014/1/1} \item{time_numeric}{Time in
  hour} \item{Temp}{Temperature} \item{Snet}{Absorbed short-wave radiation}
  \item{VPD}{Vapour pressure deficit} \item{WS}{Windspeed}}
}
\usage{
TROLLv4_dailyvar
}
\description{
Daytime variation parameters used by TROLL models from Paracou in 2014 &
2015.
}
\keyword{datasets}
