% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv4_input.R
\docType{data}
\name{TROLLv4_input}
\alias{TROLLv4_input}
\title{TROLL global parameters}
\format{
A data frame with 61 rows and 3 variables: \describe{
  \item{param}{global parameter} \item{value}{value of the parameter}
  \item{description}{description of the parameter} }
}
\usage{
TROLLv4_input
}
\description{
Global parameters definition used by TROLL model.
}
\keyword{datasets}
