% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv4_pedology.R
\docType{data}
\name{TROLLv4_pedology}
\alias{TROLLv4_pedology}
\title{TROLL pedology parameters}
\format{
A data frame with 5 rows and 8 variables: \describe{
\item{layer_thickness}{layer thickness in m}
\item{proportion_Silt}{proportion of silt in percent}
\item{proportion_Clay}{proportion of clay in percent}
\item{proportion_Sand}{proportion of sand in percent} \item{SOC}{soil
organic content} \item{DBD}{soil dry bulk density} \item{pH}{soil pH}
\item{CEC}{soil cation exchange capacity}}
}
\usage{
TROLLv4_pedology
}
\description{
Pedology parameters used by TROLL model for 5 soil layers in French Guiana.
}
\keyword{datasets}
