% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv4_species.R
\docType{data}
\name{TROLLv4_species}
\alias{TROLLv4_species}
\title{\code{TROLL} species parameters}
\format{
A data frame with 109 rows and 11 variables: \describe{
\item{s_name}{Species name genus_species} \item{s_LMA}{leaf mass per area}
\item{s_Nmass}{leaf nitrogen mass} \item{s_Pmass}{leaf phosphorus mass}
\item{s_wsg}{wood specific gravity} \item{s_dbhmax}{maximum diameter}
\item{s_hmax}{maximum height} \item{s_ah}{height-diameter allometry
coefficient} \item{s_regionalfreq}{regional frequency} \item{s_tlp}{Turgor
loss point} \item{s_leafarea}{leaf area}}
}
\usage{
TROLLv4_species
}
\description{
Functional traits used by \code{TROLL} model for 109 species in French Guiana.
}
\seealso{
\code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}
}
\keyword{datasets}
