% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.trollsim.R
\name{update_parameters}
\alias{update_parameters}
\alias{update_parameters,trollsim-method}
\title{Function to update parameters from a TROLL outputs for a next simulation.}
\usage{
update_parameters(sim, ...)

\S4method{update_parameters}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{...}{parameters to update and their values..}
}
\value{
data.frame
}
\description{
Function to update parameters from a TROLL outputs for a next simulation.
}
\examples{

data("TROLLv3_output")
head(update_parameters(TROLLv3_output, iters = 10))

}
