% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualize.R
\name{virtualize}
\alias{virtualize}
\title{Virtualize}
\usage{
virtualize(forest = getOption("RconTroll.forest"),
  path = getOption("RconTroll.path"), overwrite = FALSE,
  species = read.table(getOption("RconTroll.path"), header = TRUE, dec = ".",
  sep = "", row.names = 1), missing = getOption("RconTroll.missing"),
  spcorrect = TRUE,
  data = read.csv("~/Documents/BIOGET/Projet/Data/paracou data - corrected DBH/paracou_p1_15.csv"),
  plotcol = "n_parcelle", xcol = "Xutm", ycol = "Yutm",
  censuscol = "campagne", spcol = c("Genre", "espece"), dbhcol = "dbh_c",
  alivecol = "status", NV = 1, NH = 1, plots = c(1), year = 1984,
  alive = 1)
}
\arguments{
\item{forest}{int. path to the file to create}

\item{path}{char. working directory}

\item{overwrite}{logical. allow to overwrite existing input file}

\item{species}{df. species trait values (can be open from external file)}

\item{missing}{(logical, path). missing species should be save and where ?}

\item{spcorrect}{logical. should species will try to be corrected}

\item{data}{df. censuses data (can be open from external file)}

\item{plotcol}{char. name of the column containing plot information}

\item{xcol}{char. name of the column containing longitude information}

\item{ycol}{char. name of the column containing latitude information}

\item{censuscol}{char. name of the column containing census year information}

\item{spcol}{char. name of the column(s) containing secies information (can be a vector if species and genus are split)}

\item{dbhcol}{char. name of the column containing dbh information}

\item{alivecol}{char. name of the column containing alive information}

\item{NV}{int. vertical number of cells (nb per m)}

\item{NH}{int. horizontal number of cells (nb per m)}

\item{year}{int. year of the census to intialize on}

\item{alive}{int/char/logical. code corresponding to alive tree}

\item{plot}{int/char. name or number of the plot to initialize on}
}
\value{
virtual forest file for the TROLL program (forest.txt)
}
\description{
Virtualize a forest survey in forest text file to initialize TROLL model
}
\examples{

}
