% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualizeFromOutput.R
\name{virtualizeFromOutput}
\alias{virtualizeFromOutput}
\title{Virtualize from output}
\usage{
virtualizeFromOutput(x, forest = getOption("RconTroll.forest"),
  path = getOption("RconTroll.path"), overwrite = TRUE)
}
\arguments{
\item{x}{TROLLoutput}

\item{forest}{char. name of the file to write}

\item{path}{char. path to the working directory}

\item{overwrite}{logical. allow to overwrite existing input file}
}
\value{
virtual forest file for the TROLL program (forest.txt)
}
\description{
Virtualize  a forest from a TROLL model output
}
\examples{

}
