% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.TROLLsimstack.R
\docType{methods}
\name{aggregate.TROLLsimstack}
\alias{aggregate.TROLLsimstack}
\title{aggregate}
\usage{
\S4method{aggregate}{TROLLsimstack}(x, add = FALSE, force = FALSE, ...)
}
\arguments{
\item{x}{TROLLsimstack. TROLLsimstack to aggregate}

\item{add}{logical. should the results be added to the existing object}

\item{force}{logical. should we force the aggregation (can cause error)}

\item{...}{unused argument}
}
\value{
aggregated TROLL simstack
}
\description{
Splits the data into subsets, computes summary statistics for each, and
returns the result in a convenient form.
}
\examples{

}
