% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.troll.R
\name{autoplot,trollsim-method}
\alias{autoplot,trollsim-method}
\title{Plot TROLL simulation}
\usage{
\S4method{autoplot}{trollsim}(
  object,
  what = "ecosystem",
  variables = NULL,
  iter = NULL,
  species = NULL
)
}
\arguments{
\item{object}{TROLL simulation.}

\item{what}{char. What to plot: "forest", "ecosystem", or "species".}

\item{variables}{char. Wich variable(s) to plot (ecosystem or species).}

\item{iter}{char. Which iteration(s) to plot.}

\item{species}{char. Which species to plot when using the species table.}
}
\value{
ggplot2 object
}
\description{
This is a method to plot TROLL simulations, including either temporal
trajectories of whole ecosystem or species metrics, or the initial or final
pattern observed in the forest community. Metrics includes abundances of
individuals above 1cm (N), above 10cm (N10), and above 30cm (N30),
aboveground biomass (AGB), basal area of individuals above 1cm (BA), and
above 10cm (BA10), gross primary production (GPP), net primary production
(NPP), respiration of day (Rday), night (Rnight) and stem (Rstem), and
litterfall.
}
\examples{
\dontrun{
autoplot(sim)
}

}
