% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_climate.R
\name{generate_climate}
\alias{generate_climate}
\title{Generate climate dataset}
\usage{
generate_climate(hourly_data, daytime_start = 7, daytime_end = 19)
}
\arguments{
\item{hourly_data}{data.frame. Half-hourly climatic data formatted for TROLL.}

\item{daytime_start}{int. Daytime starting hour to compute nigh and day
variables (default 7).}

\item{daytime_end}{int. Daytime ending hour to compute nigh and day variables
(default 19).}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with climatic day variation (night temperature and
rainfall).
}
\description{
Description
}
