% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Generate parameters}
\usage{
generate_parameters(
  cols = 200,
  rows = 200,
  HEIGHT = 70,
  length_dcell = 25,
  nbiter = 10,
  iterperyear = 365,
  NV = 1,
  NH = 1,
  nbout = 4,
  nbspp = 45,
  SWtoPPFD = 2.27,
  p_nonvert = 0.05,
  klight = 0.63,
  phi = 0.093,
  absorptance_leaves = 0.9,
  theta = 0.7,
  g1 = 3.77,
  vC = 0.021,
  DBH0 = 0.005,
  H0 = 0.95,
  CR_min = 0.3,
  CR_a = 2.13,
  CR_b = 0.63,
  CD_a = 0,
  CD_b = 0.2,
  CD0 = 0.3,
  shape_crown = 0.72,
  dens = 1,
  fallocwood = 0.35,
  falloccanopy = 0.25,
  Cseedrain = 50000,
  nbs0 = 10,
  sigma_height = 0.19,
  sigma_CR = 0.29,
  sigma_CD = 0,
  sigma_P = 0.24,
  sigma_N = 0.12,
  sigma_LMA = 0.24,
  sigma_wsg = 0.06,
  sigma_dbhmax = 0.05,
  corr_CR_height = 0,
  corr_N_P = 0.65,
  corr_N_LMA = -0.43,
  corr_P_LMA = -0.39,
  leafdem_resolution = 30,
  p_tfsecondary = 1,
  hurt_decay = 0,
  crown_gap_fraction = 0.15,
  m = 0.013,
  m1 = 0.013,
  Cair = 400,
  LL_parameterization = 1,
  LA_regulation = 2,
  sapwood = 1,
  seedsadditional = 0
)
}
\arguments{
\item{cols}{num.}

\item{rows}{num.}

\item{HEIGHT}{num.}

\item{length_dcell}{num.}

\item{nbiter}{num.}

\item{iterperyear}{num.}

\item{NV}{num.}

\item{NH}{num.}

\item{nbout}{num.}

\item{nbspp}{num.}

\item{SWtoPPFD}{num.}

\item{p_nonvert}{num.}

\item{klight}{num.}

\item{phi}{num.}

\item{absorptance_leaves}{num.}

\item{theta}{num.}

\item{g1}{num.}

\item{vC}{num.}

\item{DBH0}{num.}

\item{H0}{num.}

\item{CR_min}{num.}

\item{CR_a}{num.}

\item{CR_b}{num.}

\item{CD_a}{num.}

\item{CD_b}{num.}

\item{CD0}{num.}

\item{shape_crown}{num.}

\item{dens}{num.}

\item{fallocwood}{num.}

\item{falloccanopy}{num.}

\item{Cseedrain}{num.}

\item{nbs0}{num.}

\item{sigma_height}{num.}

\item{sigma_CR}{num.}

\item{sigma_CD}{num.}

\item{sigma_P}{num.}

\item{sigma_N}{num.}

\item{sigma_LMA}{num.}

\item{sigma_wsg}{num.}

\item{sigma_dbhmax}{num.}

\item{corr_CR_height}{num.}

\item{corr_N_P}{num.}

\item{corr_N_LMA}{num.}

\item{corr_P_LMA}{num.}

\item{leafdem_resolution}{num.}

\item{p_tfsecondary}{num.}

\item{hurt_decay}{num.}

\item{crown_gap_fraction}{num.}

\item{m}{num.}

\item{m1}{num.}

\item{Cair}{num.}

\item{LL_parameterization}{num.}

\item{LA_regulation}{num.}

\item{sapwood}{num.}

\item{seedsadditional}{num.}
}
\value{
a data frame of global parameters
}
\description{
Generate parameters
}
\examples{
NA

}
