% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chm.trollsim.R
\name{get_chm}
\alias{get_chm}
\alias{get_chm,trollsim-method}
\title{Function to get a canopy height model from TROLL outputs with lidar option.}
\usage{
get_chm(sim, method = "smoothed", ...)

\S4method{get_chm}{trollsim}(sim, method = "smoothed", ...)
}
\arguments{
\item{sim}{trollsim.}

\item{method}{char. method to extract the ccanopy hieght model from the point
cloud in las, either 'filled' (replacing NA by 0) or 'smoothed' (local
means, default value).}

\item{...}{unused argument.}
}
\value{
data.frame
}
\description{
Function to get a canopy height model from TROLL outputs with lidar option.
}
