% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest.trollsim.R
\name{get_forest}
\alias{get_forest}
\alias{get_forest,trollsim-method}
\title{Function to get a forest inventory from TROLL outputs.}
\usage{
get_forest(sim, ...)

\S4method{get_forest}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{...}{unused argument.}
}
\value{
data.frame
}
\description{
Function to get a forest inventory from TROLL outputs.
}
\examples{

data("TROLLv3_output")
head(get_forest(TROLLv3_output))

}
