% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.trollsim.R
\name{get_log}
\alias{get_log}
\alias{get_log,trollsim-method}
\title{Function to get the log from TROLL outputs.}
\usage{
get_log(sim, ...)

\S4method{get_log}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{...}{unused argument.}
}
\value{
the log in the console
}
\description{
Function to get the log from TROLL outputs.
}
\examples{

get_log(TROLLv3_output)

}
