% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Function to load TROLL output}
\usage{
load_output(name, path, type)
}
\arguments{
\item{name}{char. name given to the model output}

\item{path}{char. path where the model is saved}

\item{type}{char. type of simulation (full, reduce or abc)}
}
\value{
an S4 \linkS4class{trollsim} class object
}
\description{
Function to load TROLL output
}
\examples{

NA
}
