% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Function to load outputs from a TROLL simulation.}
\usage{
load_output(name, path, thin = NULL, inmemory = TRUE)
}
\arguments{
\item{name}{char. Name given to the model output.}

\item{path}{char. Path where the model is saved.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce output size, default is NULL and corresponds to no
thinning.}

\item{inmemory}{bool. Load outputs in memory.}
}
\value{
An S4 \linkS4class{trollsim} class object.
}
\description{
Function to load outputs from a TROLL simulation.
}
\examples{

\dontrun{
 load_output("test", "./")
}

}
