% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_output}
\alias{load_output}
\title{Function to load TROLL outputs}
\usage{
load_output(name, path, thin = NULL)
}
\arguments{
\item{name}{char. Name given to the model output.}

\item{path}{char. Path where the model is saved.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce outputs size, default is NULL and corresponds to no
thinning.}
}
\value{
an S4 \linkS4class{trollsim} class object
}
\description{
Function to load TROLL outputs
}
\examples{

\dontrun{
 load_output("test", "./")
}

}
