% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sim.trollsim.R
\name{load_sim}
\alias{load_sim}
\alias{load_sim,trollsim-method}
\title{Function to load in memory outputs from a TROLL simulation.}
\usage{
load_sim(sim, ...)

\S4method{load_sim}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{...}{unused argument.}
}
\value{
An S4 \linkS4class{trollsim} class object.
}
\description{
Function to load in memory outputs from a TROLL simulation.
}
