% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stack.R
\name{load_stack}
\alias{load_stack}
\title{Function to load a stack of outputs from TROLL simulations.}
\usage{
load_stack(name, path, thin = NULL)
}
\arguments{
\item{name}{char. Name given to the stack output.}

\item{path}{char. Path where the stack is saved.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce output size, default is NULL and corresponds to no
thinning.}
}
\value{
An S4 \linkS4class{trollsim} class object.
}
\description{
Function to load a stack of outputs from TROLL simulations.
}
\examples{

\dontrun{
 load_stack("test", "./")
}

}
