% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStack.R
\name{loadStack}
\alias{loadStack}
\title{load simulation stack}
\usage{
loadStack(path, ignore = NULL)
}
\arguments{
\item{path}{char. path containing all models folders}

\item{ignore}{char. names of folder or files to ignore in path that are not
model folders}
}
\value{
an S4 \linkS4class{TROLLsimstack} class object
}
\description{
load simulation stack
}
\examples{
NA

}
