% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{option.rcontroll}
\alias{option.rcontroll}
\title{Options}
\arguments{
\item{rcontroll.tmp}{char. Path to temporary files folder used by \code{\link[=troll]{troll()}}
and \code{\link[=stack]{stack()}}.}

\item{rcontroll.troll}{char. \code{TROLL}version number accessible with
\code{\link[=TROLL.version]{TROLL.version()}}.}
}
\description{
\code{rcontroll} package global options including temporary files location and
`TROLL version.
}
