% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{option.RconTROLL}
\alias{option.RconTROLL}
\title{Options}
\arguments{
\item{name}{char. name of simulations}

\item{src}{char. path to the TROLL C++ main code (default: package tools folder)}

\item{app}{char. path to the TROLL executable (default: package tools folder)}

\item{init}{char. path to the input file template (default: package extdata folder)}

\item{species}{char. path to the species input data (default: package extdata folder)}

\item{forest}{char. name of the forest inventory input file (default: forest.txt)}
}
\description{
RconTROLL package global options
}
