% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{option.RconTroll}
\alias{option.RconTroll}
\title{Options}
\arguments{
\item{path}{char. working directory path}

\item{src}{char. path to the TROLL C++ main code (default: package tools folder)}

\item{app}{char. name of the TROLL executable (default: TROLL.out)}

\item{init}{char. name of the input file (default: input.txt)}

\item{forest}{char. name of the virtual forest input file (default: forest.txt)}

\item{species}{char. path to the species input data (default: package extdata folder)}

\item{missing}{char. name of missing species file (default: missing.txt)}

\item{output}{char. name of the output folder (default: OUTPUT)}
}
\description{
RconTROLL package global options
}
