% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trollsim.R
\name{print.trollsim}
\alias{print.trollsim}
\alias{print,trollsim-method}
\alias{show,trollsim-method}
\alias{summary,trollsim-method}
\title{Print a summary}
\usage{
\S4method{print}{trollsim}(x, ...)

\S4method{show}{trollsim}(object)

\S4method{summary}{trollsim}(object, ...)
}
\arguments{
\item{x}{trollsim or trollstack.}

\item{...}{unused argument.}

\item{object}{trollsim  or trollstack.}
}
\value{
Print or show in console.
}
\description{
\code{print()} prints a summary of TROLL simulation or stack of simulations
outputs.
}
\examples{

data("TROLLv3_output")
print(TROLLv3_output)

}
\seealso{
\code{\link[=troll]{troll()}}, \code{\link[=stack]{stack()}}, \code{\link[=trollsim]{trollsim()}}, \code{\link[=trollstack]{trollstack()}}
}
