% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TROLLsim.R
\docType{methods}
\name{print.TROLLsim}
\alias{print.TROLLsim}
\title{Function to print TROLL outputs.}
\usage{
\S4method{print}{TROLLsim}(x, ...)
}
\arguments{
\item{x}{TROLLsim}
}
\value{
Print in console
}
\description{
Function to print TROLL outputs.
}
\examples{

}
