% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trollsim.R
\name{print.trollsim}
\alias{print.trollsim}
\alias{print,trollsim-method}
\alias{show,trollsim-method}
\title{Function to print or show TROLL outputs.}
\usage{
\S4method{print}{trollsim}(x, ...)

\S4method{show}{trollsim}(object)
}
\arguments{
\item{x}{trollsim or trollstack}

\item{...}{unused argument}

\item{object}{trollsim  or trollstack}
}
\value{
Print or show in console
}
\description{
Function to print or show TROLL outputs.
}
\examples{
NA
}
