% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcontroll.R
\docType{package}
\name{rcontroll}
\alias{rcontroll}
\title{rcontroll}
\description{
Function and code to manage and use indivdual-based and spatially-explicit
forest growth model TROLL.
}
\section{Section functions}{
 \describe{
To define.
}
}

