% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_era.R
\name{request_era}
\alias{request_era}
\title{Request ERA5 data for TROLL}
\usage{
request_era(
  variables = c("u10", "v10", "d2m", "t2m", "ssr", "sp", "tp"),
  years = 2004:2005,
  months = 1:12,
  days = 1:31,
  hours = 0:23,
  prefix = "ERA5land_Paracou",
  xutm = 5.26724134423233,
  yutm = -52.924368025558,
  format = "netcdf"
)
}
\arguments{
\item{variables}{char. Requested variable among "u10", "v10", "d2m", "t2m",
"ssr", "sp", and "tp".}

\item{years}{int. Requested years between 1950 and today.}

\item{months}{int. Requested months between 1 and 12.}

\item{days}{int. Requested months between 1 and 31.}

\item{hours}{int. Requested hours between 0 and 23.}

\item{prefix}{char. Prefix used to name the file.}

\item{xutm}{int. Requested site longitude in UTM.}

\item{yutm}{int. Requested site latitude in UTM.}

\item{format}{char. Requested format, default netcdf.}
}
\value{
ERA request for \code{ecmwfr} as a list of attributes.
}
\description{
Description.
}
