% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Stack}
\usage{
stack(
  name = NULL,
  simulations,
  path = NULL,
  full = TRUE,
  abc = FALSE,
  random = TRUE,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  cores = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{name}{char. stack name (if NULL timestamp)}

\item{simulations}{char. simulations names}

\item{path}{char. path to the stack (tmp if NULL)}

\item{full}{bool. TROLL with full outputs, if not reduced outputs (default
TRUE)}

\item{abc}{bool. TROLL with abc outputs, forcing reduced outputs (default
FALSE)}

\item{random}{bool. TROLL with random outputs, if not the seed is fixed
(default TRUE)}

\item{global}{df. global parameters, add a simulation column for repeated
simulations}

\item{species}{df. species parameters, add a simulation column for repeated
simulations}

\item{climate}{df. climate parameters, add a simulation column for repeated
simulations}

\item{daily}{df. daily variation parameters, add a simulation column for
repeated simulations}

\item{forest}{df. TROLL with forest input, if null start from an empty grid,
add a simulation column for repeated simulations (default NULL)}

\item{cores}{int. number of cores for parallelization, if NULL available
cores - 1 (default NULL)}

\item{overwrite}{bool. overwrite previous outputs}
}
\value{
trollstack
}
\description{
Run a TROLL stack
}
\examples{
\dontrun{
example
}

}
