% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Stack}
\usage{
stack(
  name = NULL,
  simulations,
  path = NULL,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  verbose = TRUE,
  cores = NULL,
  overwrite = TRUE,
  thin = NULL
)
}
\arguments{
\item{name}{char. Stack name (if NULL timestamp).}

\item{simulations}{char. Simulations names.}

\item{path}{char. Path to save the simulations outputs, the default is null
corresponding to a simulation in memory without saved intermediary files.}

\item{global}{df. Global parameters.}

\item{species}{df. Species parameters.}

\item{climate}{df. Climate parameters.}

\item{daily}{df. Daily variation parameters.}

\item{forest}{df. TROLL with forest input, if null start from an empty grid
(default NULL).}

\item{verbose}{bool. Show TROLL outputs in the console.}

\item{cores}{int. Number of cores for parallelization, if NULL available
cores - 1 (default NULL).}

\item{overwrite}{bool. Overwrite previous outputs.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce outputs size, default is NULL and corresponds to no
thinning.}
}
\value{
trollstack
}
\description{
Run a TROLL stack.
}
\examples{
\dontrun{
stack(name = "test3", 
simulations = c("seed50000", "seed500"),
global = TROLLv3_input_stack,
species = TROLLv3_species,
climate = TROLLv3_climatedaytime12,
daily = TROLLv3_daytimevar,
verbose = F,
cores = 2,
thin = c(1,5,10))
}

}
