% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trollsim.R
\name{summary.trollsim}
\alias{summary.trollsim}
\alias{summary,trollsim-method}
\title{Function to provide summary of TROLL outputs}
\usage{
\S4method{summary}{trollsim}(object, ...)
}
\arguments{
\item{object}{trollsim  or trollstack.}

\item{...}{unused argument.}
}
\value{
Print in console.
}
\description{
Function to provide summary of TROLL outputs
}
\examples{
NA
}
