% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TROLLsim.R
\docType{methods}
\name{summary.TROLLsim}
\alias{summary.TROLLsim}
\alias{summary,TROLLsim-method}
\alias{summary,TROLLsimstack-method}
\title{Function to provide summary of TROLL outputs}
\usage{
\S4method{summary}{TROLLsim}(object, ...)

\S4method{summary}{TROLLsimstack}(object, ...)
}
\arguments{
\item{object}{TROLLsim or TROLLsimstack}

\item{...}{unused argument}
}
\value{
Print in console
}
\description{
Function to provide summary of TROLL outputs
}
\examples{
NA

}
