% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TROLLsimstack.R
\docType{methods}
\name{summary.TROLLsimstack}
\alias{summary.TROLLsimstack}
\title{Function to provide summary of TROLL outputs}
\usage{
\S4method{summary}{TROLLsimstack}(object, ...)
}
\arguments{
\item{x}{TROLLsimstack}
}
\value{
Print in console
}
\description{
Function to provide summary of TROLL outputs
}
\examples{

}
