% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{TROLL}
\usage{
troll(
  name = NULL,
  path = NULL,
  troll = NULL,
  general = data.frame(),
  species = data.frame(),
  climate = data.frame(),
  daily = data.frame(),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{name}{char. model name}

\item{path}{char. path to the simulation (tmp if NULL)}

\item{troll}{chr. path to troll executable (default in the package)}

\item{general}{df. general parameters}

\item{species}{df. species parameters}

\item{climate}{df. climate parameters}

\item{daily}{df. daily variation parameters}

\item{overwrite}{bool. overwrite previous outputs}

\item{verbose}{bool. display models information}
}
\value{
simulation outputs in the path folder
}
\description{
Run a TROLL simulation
}
\examples{
\dontrun{
example
}

}
