% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{troll}
\usage{
troll(
  name = NULL,
  path = NULL,
  global,
  species,
  climate,
  daily,
  pedology,
  forest = NULL,
  soil = NULL,
  lidar = NULL,
  verbose = TRUE,
  load = TRUE,
  overwrite = TRUE,
  thin = NULL,
  date = NULL
)
}
\arguments{
\item{name}{char. Model name (if NULL timestamp).}

\item{path}{char. Path to save the simulation outputs, the default is null
corresponding to a simulation in memory without saved intermediary files.}

\item{global}{df. Global parameters.}

\item{species}{df. Species parameters.}

\item{climate}{df. Climate parameters.}

\item{daily}{df. Daily variation parameters.}

\item{pedology}{df. Daily variation parameters.}

\item{forest}{df. TROLL with forest input, if null starts from an empty grid
(default NULL).}

\item{soil}{df. TROLL with soil input, if null starts from an empty grid
(default NULL).}

\item{lidar}{df. Lidar simulation parameters, if null no computed (default
NULL).}

\item{verbose}{bool. Show TROLL outputs in the console.}

\item{load}{bool. TROLL outputs are loaded in R memory, if not only the path
to the outputs is kept.}

\item{overwrite}{bool. Overwrite previous outputs.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce output size, default is NULL and corresponds to no thinning.}

\item{date}{char. Starting date as YYYY/MM/DD, default NULL will result in
non-dated outputs.}
}
\value{
A trollsim object.
}
\description{
Run a TROLL simulation.
}
\examples{
\donttest{
data("TROLLv4_input")
data("TROLLv4_species")
data("TROLLv4_climate")
data("TROLLv4_dailyvar")
data("TROLLv4_pedology")
sim <- troll(
  name = "test",
  path = getwd(),
  global = generate_parameters(nbiter = 10),
  species = TROLLv4_species,
  climate = TROLLv4_climate,
  daily = TROLLv4_dailyvar,
  pedology = TROLLv4_pedology,
  load = TRUE,
  date = "2004/01/01"
)
}

}
