% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{troll}
\usage{
troll(
  name = NULL,
  path = NULL,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  verbose = TRUE,
  overwrite = TRUE,
  thin = NULL
)
}
\arguments{
\item{name}{char. Model name (if NULL timestamp).}

\item{path}{char. Path to save the simulations outputs, the default is null
corresponding to a simulation in memory without saved intermediary files.}

\item{global}{df. Global parameters.}

\item{species}{df. Species parameters.}

\item{climate}{df. Climate parameters.}

\item{daily}{df. Daily variation parameters.}

\item{forest}{df. TROLL with forest input, if null start from an empty grid
(default NULL).}

\item{verbose}{bool. Show TROLL outputs in the console.}

\item{overwrite}{bool. Overwrite previous outputs.}

\item{thin}{int. Vector of integers corresponding to the iterations to be
kept to reduce outputs size, default is NULL and corresponds to no
thinning.}
}
\value{
trollsim
}
\description{
Run a TROLL simulation.
}
\examples{
\dontrun{
sim <- troll(name = "test",
global = generate_parameters(iterperyear = 12, nbiter = 12*1),
species = TROLLv3_species,
climate = TROLLv3_climatedaytime12,
daily = TROLLv3_daytimevar)
}

}
