% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{TROLL}
\usage{
troll(
  name = NULL,
  path = NULL,
  full = TRUE,
  abc = FALSE,
  random = TRUE,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{name}{char. model name (if NULL timestamp)}

\item{path}{char. path to the simulation (tmp if NULL)}

\item{full}{bool. TROLL with full outputs, if not reduced outputs (default
TRUE)}

\item{abc}{bool. TROLL with abc outputs, forcing reduced outputs (default
FALSE)}

\item{random}{bool. TROLL with random outputs, if not the seed is fixed
(default TRUE)}

\item{global}{df. global parameters}

\item{species}{df. species parameters}

\item{climate}{df. climate parameters}

\item{daily}{df. daily variation parameters}

\item{forest}{df. TROLL with forest input, if null start from an empty grid
(default NULL)}

\item{overwrite}{bool. overwrite previous outputs}
}
\value{
trollsim
}
\description{
Run a TROLL simulation
}
\examples{
\dontrun{
example
}

}
