% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{TROLL}
\usage{
troll(
  name = NULL,
  path = NULL,
  troll_path = NULL,
  troll_exe = "TROLL_full.out",
  global = data.frame(),
  species = data.frame(),
  climate = data.frame(),
  daily = data.frame(),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{name}{char. model name}

\item{path}{char. path to the simulation (tmp if NULL)}

\item{troll_path}{chr. path to troll executable (default in the package)}

\item{troll_exe}{chr. name of troll executable (TROLL_abc.out,
TROLL_full.out, TROLL_full_forest.out, TROLL_reduced.out,
TROLL_reduced_forest.out)}

\item{global}{df. global parameters}

\item{species}{df. species parameters}

\item{climate}{df. climate parameters}

\item{daily}{df. daily variation parameters}

\item{overwrite}{bool. overwrite previous outputs}

\item{verbose}{bool. display models information}
}
\value{
simulation outputs in the path folder
}
\description{
Run a TROLL simulation
}
\examples{
\dontrun{
example
}

}
