% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{TROLL}
\usage{
troll(
  name = NULL,
  path = NULL,
  global,
  species,
  climate,
  daily,
  forest = NULL,
  verbose = TRUE,
  overwrite = TRUE,
  thin = NULL
)
}
\arguments{
\item{name}{char. model name (if NULL timestamp)}

\item{path}{char. path to the simulation (tmp if NULL)}

\item{global}{df. global parameters}

\item{species}{df. species parameters}

\item{climate}{df. climate parameters}

\item{daily}{df. daily variation parameters}

\item{forest}{df. TROLL with forest input, if null start from an empty grid
(default NULL)}

\item{verbose}{bool. show TROLL outputs in the console}

\item{overwrite}{bool. overwrite previous outputs}

\item{thin}{int. vector of integers corresponding to iterations to be kept
(default NULL)}
}
\value{
trollsim
}
\description{
Run a TROLL simulation
}
\examples{
\dontrun{
example
}

}
