% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troll.R
\name{troll}
\alias{troll}
\title{TROLL}
\usage{
troll(
  name = NULL,
  path = NULL,
  full = TRUE,
  abc = FALSE,
  global = data.frame(),
  species = data.frame(),
  climate = data.frame(),
  daily = data.frame(),
  overwrite = TRUE
)
}
\arguments{
\item{name}{char. model name}

\item{path}{char. path to the simulation (tmp if NULL)}

\item{full}{bool. TROLL with full outputs, if not reduced outputs (default
TRUE)}

\item{abc}{bool. TROLL with abc outputs, forcing reduced outputs (default
FALSE)d}

\item{global}{df. global parameters}

\item{species}{df. species parameters}

\item{climate}{df. climate parameters}

\item{daily}{df. daily variation parameters}

\item{overwrite}{bool. overwrite previous outputs}
}
\value{
simulation outputs in the path folder
}
\description{
Run a TROLL simulation
}
\examples{
\dontrun{
example
}

}
