% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLversion.R
\name{TROLL.version}
\alias{TROLL.version}
\title{Function to print TROLL version.}
\usage{
TROLL.version()
}
\value{
Print TROLL version in console.
}
\description{
Function to print TROLL version.
}
\examples{

TROLL.version()

}
