% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trollCpp}
\alias{trollCpp}
\title{TROLL C++}
\usage{
trollCpp(global_file, climate_file, species_file, day_file, output_file)
}
\arguments{
\item{global_file}{std::string. global parameters file.}

\item{climate_file}{std::string. climate file.}

\item{species_file}{std::string. species file.}

\item{day_file}{std::string. daytime file.}

\item{output_file}{std::string. output folder.}
}
\value{
int.
}
\description{
TROLL C++.
}
\examples{
trollCpp(global_file = "test/test_input_global.txt", climate_file = "test/test_input_climate.txt", species_file = "test/test_input_species.txt", day_file = "test/test_input_daily.txt", output_file = "test")

}
