% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsim.R
\docType{class}
\name{trollsim-class}
\alias{trollsim-class}
\title{An S4 class to represent TROLL simulations}
\description{
This is an S4 class to represent TROLL simulations.
}
\section{Slots}{

\describe{
\item{\code{name}}{char. model name}

\item{\code{path}}{char. path to the model}

\item{\code{inputs}}{list. model inputs (general, species, climate, daily, forest)}

\item{\code{log}}{chr. model log}
}}

