% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsim.R
\docType{class}
\name{trollsim-class}
\alias{trollsim-class}
\title{An S4 class to represent TROLL simulations}
\description{
This is an S4 class to represent TROLL simulations.
}
\section{Slots}{

\describe{
\item{\code{name}}{char. simulation name}

\item{\code{path}}{char. path to the simulation}

\item{\code{parameters}}{numeric. parameters of the simulation (general inputs, random, forest)}

\item{\code{inputs}}{list. simulation inputs (species, climate, daily, forest)}

\item{\code{log}}{chr. simulation log}

\item{\code{final_pattern}}{df. simulation final forest}
}}

