% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLsim.R
\docType{class}
\name{TROLLsim-class}
\alias{TROLLsim-class}
\title{An S4 class to represent TROLL outputs}
\description{
This is an S4 class to represent TROLL outputs.
}
\section{Slots}{

\describe{
\item{\code{name}}{char. model name}

\item{\code{path}}{char. path to the model}

\item{\code{abundances}}{list. abundances data frames}

\item{\code{agb}}{df. agb data frame}

\item{\code{ba}}{list. ba data frames}

\item{\code{dbh}}{df. dbh data frame}

\item{\code{death}}{list. death data frames}

\item{\code{disturbance}}{df. tree killed in disturbance data frame}

\item{\code{final_pattern}}{df. final pattern data frame}

\item{\code{gpp}}{df. gpp data frame}

\item{\code{info}}{list. model info}

\item{\code{litterfall}}{df. litterfall data frame}

\item{\code{npp}}{df. npp data frame}

\item{\code{par}}{df. par data frame}

\item{\code{paramspace}}{list. model space parameters}

\item{\code{ppfd0}}{df. ground level ppfd data frames}

\item{\code{R}}{list. respiration data frames}

\item{\code{sp_par}}{df. species par data frame}

\item{\code{vertd}}{df. vertd data.frame}
}}

