% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsim.R
\docType{class}
\name{trollsim}
\alias{trollsim}
\title{An S4 class to represent TROLL simulations}
\usage{
trollsim(
  name = character(),
  path = character(),
  abundances = list(),
  agb = data.frame(),
  ba = list(),
  death = list(),
  final_pattern = data.frame(),
  gpp = data.frame(),
  info = character(),
  log = character(),
  litterfall = data.frame(),
  npp = data.frame(),
  inputs = list(),
  ppfd0 = data.frame(),
  r = list(),
  vertd = data.frame()
)
}
\description{
This is an S4 class to represent TROLL simulations.
}
\section{Slots}{

\describe{
\item{\code{name}}{char. model name}

\item{\code{path}}{char. path to the model}

\item{\code{abundances}}{list. abundances data frames}

\item{\code{agb}}{df. agb data frame}

\item{\code{ba}}{list. ba data frames}

\item{\code{death}}{list. death data frames}

\item{\code{final_pattern}}{df. final pattern data frame}

\item{\code{gpp}}{df. gpp data frame}

\item{\code{info}}{chr. model info}

\item{\code{log}}{chr. model log}

\item{\code{litterfall}}{df. litterfall data frame}

\item{\code{npp}}{df. npp data frame}

\item{\code{inputs}}{list. model inputs (general, species, climate, daily, soil)}

\item{\code{ppfd0}}{df. ground level ppfd data frames}

\item{\code{R}}{list. respiration data frames}

\item{\code{vertd}}{df. vertd data.frame}
}}

