% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsim.R
\name{trollsim}
\alias{trollsim}
\title{An S4 class to represent TROLL simulations}
\usage{
trollsim(
  name = character(),
  path = character(),
  parameters = numeric(),
  inputs = list(),
  log = character(),
  final_pattern = data.frame()
)
}
\arguments{
\item{name}{char. simulation name}

\item{path}{char. path to the simulation}

\item{parameters}{char. parameters of the simulation (general inputs, random, forest)}

\item{inputs}{numeric. simulation inputs (species, climate, daily, forest)}

\item{log}{chr. simulation log}

\item{final_pattern}{df. simulation final forest}
}
\description{
This is an S4 class to represent TROLL simulations.
}
