% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollstack.R
\name{trollstack}
\alias{trollstack}
\title{An S4 class to represent TROLL stack}
\usage{
trollstack(
  name = character(),
  path = character(),
  parameters = numeric(),
  inputs = list(),
  log = character(),
  final_pattern = data.frame(),
  outputs = data.frame()
)
}
\arguments{
\item{name}{char. Simulation name.}

\item{path}{char. Path to the simulation.}

\item{parameters}{numeric. Parameters of the simulation (general inputs).}

\item{inputs}{list. Simulation inputs (species, climate, daily, forest).}

\item{log}{chr. Simulation log.}

\item{final_pattern}{df. Simulation final forest.}

\item{outputs}{df. Species and ecosystem metrics (reduced or full).}
}
\description{
This is an S4 class to represent TROLL stack
}
