% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_daytimevar.R
\docType{data}
\name{TROLLv3_daytimevar}
\alias{TROLLv3_daytimevar}
\title{TROLL daytime variation parameters}
\format{
A data frame with 24 rows and 5 variables: \describe{
  \item{starttime}{?} \item{endtime}{?} \item{vardaytime_light}{daytime
  variation in light} \item{vardaytime_vpd}{daytime variation in vapor
  pressure deficit} \item{vardaytime_T}{daytime variation in temperature}}
}
\usage{
TROLLv3_daytimevar
}
\description{
Daytime variation parameters used by TROLL model.
}
\keyword{datasets}
