% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_output.R
\docType{data}
\name{TROLLv3_output}
\alias{TROLLv3_output}
\title{TROLL output}
\format{
A trollsim object.
}
\usage{
TROLLv3_output
}
\description{
TROLL outputs from a 500 years simulations with default parameters using
TROLLv3_species,  TROLLv3_climatedaytime12, and TROLLv3_daytimevar to be used
in tests and examples.
}
\keyword{datasets}
