% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_output.R
\docType{data}
\name{TROLLv3_output}
\alias{TROLLv3_output}
\title{TROLL output}
\format{
A trollsim object.
}
\usage{
TROLLv3_output
}
\description{
TROLL outputs from a 100-year simulation on a 100x100 grid with the other
default parameters and using TROLLv3_species, TROLLv3_climatedaytime12, and
TROLLv3_daytimevar for use in tests and examples. Ecosystem level output has
been thinned and species output has been removed to save disk space.
}
\keyword{datasets}
