% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_pointcloud.R
\docType{data}
\name{TROLLv3_pointcloud}
\alias{TROLLv3_pointcloud}
\title{\code{TROLL} lidar parameters}
\format{
A data frame with 5 rows and 3 variables: \describe{
\item{param}{lidar parameter} \item{value}{value of the parameter}
\item{description}{description of the parameter} }
}
\usage{
TROLLv3_pointcloud
}
\description{
Lidar parameters definition used by \code{TROLL} for lidar simulations.
}
\seealso{
\code{\link[=generate_lidar]{generate_lidar()}}
}
\keyword{datasets}
