% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_species.R
\docType{data}
\name{TROLLv3_species}
\alias{TROLLv3_species}
\title{TROLL species parameters}
\format{
A data frame with 45 rows and 12 variables: \describe{
  \item{s_name}{species name genus_species} \item{s_LMA}{leaf mass per area}
  \item{s_Nmass}{leaf nitrogen mass} \item{s_Pmass}{leaf phosphorous mass}
  \item{s_wsg}{wood specific gravity} \item{s_dbhmax}{maximum diameter}
  \item{s_hmax}{maximum height} \item{s_ah}{height-diameter allometry
  coefficient} \item{s_regionalfreq}{regional frequency} \item{s_tlp}{Turgor
  loss point} \item{s_drymass}{leaf dry mass} }
}
\usage{
TROLLv3_species
}
\description{
Functional traits used by TROLL model for every species.
}
\keyword{datasets}
