% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.trollsim.R
\name{update_parameters}
\alias{update_parameters}
\alias{update_parameters,trollsim-method}
\title{Function to updae parameters from a TROLL outputs for a next simlation.}
\usage{
update_parameters(sim, ...)

\S4method{update_parameters}{trollsim}(sim, ...)
}
\arguments{
\item{sim}{trollsim.}

\item{...}{parameters to update and their values..}
}
\value{
data.frame
}
\description{
Function to updae parameters from a TROLL outputs for a next simlation.
}
\examples{

data("TROLLv3_output")
head(get_forest(TROLLv3_output))

}
